+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 25.3.0
+
+* [swift] Adds equality methods to generated data classes.
+* [dart, kotlin] Shortens equality methods.
+
 ## 25.2.0
 
 * [kotlin] Adds equality methods to generated data classes.